Controls4U: Now runs on 64 bits too.
git-svn-id: svn://ultimatepp.org/upp/trunk@5764 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ bazaar/Controls4U/ActiveX.cpp
@@ -28,7 +28,7 @@ String BSTRGet(BSTR &bstr) {
 	if (!(buffer = (char *)GlobalAlloc(GMEM_FIXED, sizeof(wchar_t) * size)))
 		return Null;
 	
-	int i = wcstombs(buffer, bstr, size);
+	size_t i = wcstombs(buffer, bstr, size);
 	buffer[i] = 0;
 	
 	String ret = buffer;
